// src/pages/loan/scoring/Scoring-type.ts
export type FormScoring = {
  "gender": string;
  "maritalStatus": string;
  "dependentAmount": number;
  "passportIssueDate": string;
  "passportIssueBranch": string;
  "employmentStatus": string;
  "employerINN": number;
  "salary": number;
  "position": string;
  "workExperienceTotal": number;
  "workExperienceCurrent": number;
}